+++ TDEngine2/include/core/CBasePluginManager.h
@@ -95,6 +95,14 @@ namespace TDEngine2
 			TDE2_API E_RESULT_CODE RegisterECSComponents(TPtr<IWorld> pWorld) override;
 
 			/*!
+				\brief The method allows plugins to initialize and register ECS systems
+
+				\return RC_OK if everything went ok, or some other code, which describes an error
+			*/
+
+			TDE2_API E_RESULT_CODE RegisterECSSystems(TPtr<IWorld> pWorld) override;
+
+			/*!
 				\brief The method returns a type of the subsystem
 
 				\return A type, which is represented with E_ENGINE_SUBSYSTEM_TYPE's value