+++ CHANGELOG.md
@@ -34,9 +34,11 @@ should be wrapped up with a deffered action's call.
 
 - A new type of 3d physics collider **CCapsuleCollisionObject3D** was added.
 
+### Changed
+
 - 3D physics implementation via Bullet physics engine was moved into a separate plugin.
 
-### Changed
+- The API of **IECSPlugin** was changed OnRegister methods was splitted into OnRegisterComponents and OnRegisterSystems.
 
 ### Fixed